PC-513
What is the problem to fix?

    To reproduce the problem you should follow these steps:
    Add the 2 portlet wars in the attached file to the server and run the server.
    Login
    Groups->administration->application registry and auto import.
    Site Editor->add Page Wizard create a new page and add to it one or two instances of TODOseguro portlet in addition to TODO Struts portlet and save the page.
    Navigate to the page newly created and for example push a button in the TODOseguro Portlet.
    The expected result is that only TODOseguro Portlet will respond to button click in her way.

But actually, you will notice that TODO Struts is acting like responding to the request sent to TODOseguro Portlet. So, the 2 portlets have the same behiavor despite the request is sent only to TODOseguro portlet.

This problem is noticed also if we replace the TODOseguro portlet with any other portlet.

The attached files.zip contains the TODOseguro portlet(cap5-lab3b.war),the TODO Struts portlet(cap7Struts-lab2.war) and a video showing the problem(out.ogv).

How is the problem fixed?
    Added extra checking that incoming parameter is intended to this portlet window and not belongs to other portlets.